move gini coefficient from eval to eval_metrics
