+++ eval_metrics.py
@@ -41,3 +41,45 @@ def psp(
         mpsp += 1.0 / item_propensities[item_id]
 
     return upsp / mpsp
+
+class GiniCoefficient:
+    """
+    A class to calculate the Gini coefficient, a measure of income inequality.
+    The Gini coefficient ranges from 0 (perfect equality) to 1 (perfect inequality).
+    """
+
+    def gini_coefficient(self, values):
+        """
+        Compute the Gini coefficient of array of values.
+        For a frequency vector, G = sum_i sum_j |x_i - x_j| / (2 * n^2 * mu)
+        """
+        print(f"[GINI] Computing Gini coefficient for {len(values)} values")
+        arr = np.array(values, dtype=float)
+        if arr.sum() == 0:
+            print("[GINI] Sum of values is 0, returning 0.0")
+            return 0.0
+        # sort and normalize
+        arr = np.sort(arr)
+        n = arr.size
+        cumvals = np.cumsum(arr)
+        mu = arr.mean()
+        # the formula simplifies to:
+        # G = (1 / (n * mu)) * ( sum_i (2*i - n - 1) * arr[i] )
+        index = np.arange(1, n + 1)
+        gini = (np.sum((2 * index - n - 1) * arr)) / (n * n * mu)
+        print(f"[GINI] Computed Gini coefficient: {gini:.4f}")
+        return gini
+
+    def calculate_list_gini(self, articles, key="category"):
+        """
+        Given a list of article dicts and a key (e.g. 'category'), compute the
+        Gini coefficient over the frequency distribution of that key.
+        """
+        print(f"[GINI] Calculating Gini for {len(articles)} articles using key '{key}'")
+        # count frequencies
+        freqs = {}
+        for art in articles:
+            val = art.get(key, None) or "UNKNOWN"
+            freqs[val] = freqs.get(val, 0) + 1
+        print(f"[GINI] Found {len(freqs)} unique {key} values")
+        return self.gini_coefficient(list(freqs.values()))
